requirements.txt
pillow
pillow-avif-plugin
pyinstaller
tkinterdnd2